requirements.txt
numpy 
pandas
Django
djangorestframework
djangorestframework-simplejwt
django-cors-headers
pillow
django-filter
markdown
coreapi
drf-spectacular
python-dotenv
psycopg2-binary
sqlalchemy